+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - **fix:** adjust tests
 - **doc:** update versions
 - **fix:** linting errors
+- **fix:** refactor test mocks
 
 ### [v0.4.0] - 2016-07-16
 - **fix:** version in meta data