+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased
 - **fix:** dependency listing
+- **doc:** fix example
 
 ### [v0.5.0] - 2016-07-17
 - **doc:** add documentation