+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### unreleased
 - **fix:** dependency listing
 - **doc:** fix example
+- **doc:** fix flaws
 
 ### [v0.5.0] - 2016-07-17
 - **doc:** add documentation